Change travis build, restructure tests
