+++ .travis.yml
@@ -2,6 +2,11 @@ dist: xenial
 language: python
 
 python:
+  - "3.1"
+  - "3.2"
+  - "3.3"
+  - "3.4"
+  - "3.5"
   - "3.6"
   - "3.7"
   - "3.7.4"
@@ -9,9 +14,11 @@ python:
 install:
   - pip install -r requirements.txt
   - pip install codecov
+  - pip install coverage
 
 script:
   - python -m unittest discover
+  - coverage run --source fail2slack -m test.test_fail2slack
 
 after_success:
   - codecov
